Fix circular dependency that was causing tests to fail
